broke down code into multiple files
